==> 0_eda.ipynb <==
%config ZMQInteractiveShell.ast_node_interactivity='all'
%matplotlib inline
import warnings;warnings.filterwarnings('ignore')
import numpy as np
import pandas as pd
import lightgbm as lgb
from sklearn.preprocessing import LabelEncoder
from sklearn import metrics
import os
import matplotlib.pyplot as plt
import seaborn as sns

target = '标签'
uid = '申请编号'
columns = {}
columns['no_fea'] = [uid, target, '申请时点']
columns['01'] = [
    '贷款类型',
    '是否提供手机号',
    '是否提供电话',
    '手机号是否有效',
    '是否提供email',
    '申请人是否额外提供了文件2',
    '申请人是否额外提供了文件3',
    '申请人是否额外提供了文件4',
    '申请人是否额外提供了文件5',
    '申请人是否额外提供了文件6',
    '申请人是否额外提供了文件7',
    '申请人是否额外提供了文件8',
    '申请人是否额外提供了文件9',
    '是否有车',
    '是否有房',
    '地址是否一致标志1',
    '地址是否一致标志2',
    '地址是否一致标志3',
    '地址是否一致标志4',
    '地址是否一致标志5',
    '地址是否一致标志6']
columns['100'] = [
    '性别',
    '客户居住地评分1',
    '客户居住地评分2',
    '教育程度',
    '婚姻状态',
    '居住状态',
    '贷款申请前1小时内征信查询次数',
    '贷款申请前1天内征信查询次数',
    '贷款申请前1周内征信查询次数',
    '贷款申请前1个月内征信查询次数',
    '贷款申请前1个季度内征信查询次数',
    '贷款申请前1年内征信查询次数',
    '收入类型',
    '陪同申请人',
    '孩子个数',
    '职业',
    '家庭成员数',
    '社交圈违约信息1_1',
    '社交圈违约信息2_1',
    '有车时间',
    '单位类型']
columns['time'] = [
    '申请周内日',
    '申请时点',
    '出生日期距申请日期天数']
columns['_dt'] = [
    '最近一次换手机号码距申请日天数',
    '出生日期距申请日期天数',
    '工作日期距申请日期天数',
    '注册日期距申请日期天数',
    '身份认证日期距申请日期天数']
columns['money'] = [
    '贷款年金',
    '客户收入',
    '商品价格',
    '信用额度']

############## Functions

def ana_dtypes(t):
    if t == 'str':
        cols = types[types == 'object'].index
    else:
        cols = types[types != 'object'].index
    # Save description
    cols_desc = d[cols].describe().T
    cols_desc.to_csv(f'./data/cols_desc_{t}.csv')
    # Return columns
    return cols

# a = d11.select_dtypes(include=['int64'])

############## Load

d0 = pd.read_csv('./data/train_label.csv')
d0.shape
d0.head()

# Target Count
d0['标签'].value_counts()

d1 = pd.read_csv('./data/train_Application.csv')
d1.shape
d1.head()
d1.dtypes.value_counts()

d2 = pd.read_csv('./data/train_Personas.csv')
d2.shape
d2.dtypes.value_counts()
d2.head()


### Merge Application

d11 = pd.merge(d0, d1, how='inner', on='申请编号')
d11.shape
d11.head()

d11 = pd.merge(d11, d2, how='inner', on='申请编号')
d11.shape
d11.head()

d11.to_csv(f'./tmp/0_merge11.csv')

d11.groupby(['标签', '客户居住地评分1']).agg({'申请编号':'count', '标签':'mean' })

''' 重复？'''
d11[uid].nunique()

''' Target '''
d11[target].value_counts()

''' Dtypes '''
d11.dtypes.value_counts()

''' Description '''
d_feats = d11.drop([uid, target], axis=1)
cols_desc = d_feats.describe().T
# Unique
cols_unique = d_feats.apply(pd.Series.nunique, axis = 0)
cols_unique = pd.DataFrame(cols_unique, columns=['unique'])
cols_desc = cols_desc.join(cols_unique)

cols_desc.head()
cols_desc.to_csv(f'./tmp/0_cols_desc.csv')

############## Outlier：Get

low = .01
high = .99
quant_df = d11.drop([uid, target], axis=1).quantile([low, high]).T
quant_df.to_csv('./tmp/0_outlier.csv')

# 处理
d12 = d11.copy()

# 置空值
d12.loc[d12['孩子个数']==-1, '孩子个数'] = np.NaN
# 距申请日期>0
for col in columns['_dt']:
    d12.loc[d12[col]>0, col] = np.NaN

############## Outlier：Apply

outliers = pd.read_csv('./tmp/0_outlier1.csv', index_col=0)
outliers.shape
outliers.head()

def set_outlier(col):
    if col < col_min:
        col = col_min
    elif col > col_max:
        col = col_max
    return col

# def set_outlier(col):
#     if col < col_min or col > col_max:
#         col = np.NaN
#     return col
    
for col in outliers.index:
# for col in '出生日期距申请日期天数', '社交圈违约信息1_2':
    col_min = outliers.loc[col, 'min']
    col_max = outliers.loc[col, 'max']
    d12[col] = d12[col].apply(set_outlier)
    
d12.to_csv(f'./tmp/2_eda_outlier.csv')

############# correlations

correlations = d12.drop([uid], axis=1).corr()
correlations.to_csv('./tmp/0_correlations.csv')

d12.loc[d12['出生日期距申请日期天数'], '身份认证日期距申请日期天数'].min()

for col in ['社交圈违约信息1_2']:
    d12[col].value_counts()

############## Feature Engineering

### Correlation

def calc_poly(cols):
    degree = 2
    poly_features = d12[cols]
    poly_target = d12[target]

    # Fill Miss
    from sklearn.preprocessing import Imputer
    imputer = Imputer(strategy = 'median')
    poly_features = imputer.fit_transform(poly_features)

    # Polynomial
    from sklearn.preprocessing import PolynomialFeatures
    poly_transformer = PolynomialFeatures(degree = degree, include_bias=False) # interaction_only
    poly_transformer.fit(poly_features)
    poly_features = poly_transformer.transform(poly_features)
    poly_features = pd.DataFrame(poly_features, 
                                 columns = poly_transformer.get_feature_names(cols))
    return poly_features


cols = ['出生日期距申请日期天数',
        '工作日期距申请日期天数',
        '最近一次换手机号码距申请日天数',
        '客户居住地评分2',
        '客户居住地评分1',
        '教育程度',
        '性别',
        '身份认证日期距申请日期天数',
        '地址是否一致标志6',
        '收入类型',
        '是否提供手机号',
        '地址是否一致标志2',
        '注册日期距申请日期天数',
        '有车时间']
#         '居住状态',
#         '地址是否一致标志4',
#         '是否提供电话',
#         '职业',
#         '社交圈违约信息2_1',
#         '贷款申请前1年内征信查询次数',
#         '社交圈违约信息1_1']
poly_features = calc_poly(cols)
poly_features.shape
poly_features.head()

d13 = pd.concat([d12, poly_features.drop(cols, axis=1)], axis=1)
d13.shape
d13.head()

d13.to_csv(f'./tmp/3_poly.csv')

### 导入d3历史申请表数据 ###
d3 = pd.read_csv('./data/train_History_Application.csv')
d3.shape
d3.dtypes
d3.head()

### 特征分类 ###
target = '标签'
uid = '申请编号'
hid = '历史申请编号'
col_cat = {}
col_cat['no_fea'] = [uid, target, hid]
# d3_feats = d3.drop([uid, hid], axis=1)

# 金额型
col_cat['curr'] = [
    '贷款年金',
    '申请额度',
    '信用额度',
    '首付金额',
    '商品价格']

# 类别型
col_cat['cat'] = [
    '贷款类型',
    '贷款用途',
    '合同状态',
    '付款方式',
    '拒绝原因',
    '陪同申请人',
    '客户类型',
    '商品类别',
    '组合类型',
    '产品类型',
    '获客渠道',
    '销售区域',
    '行业',
    '收益级别',
    '产品组合']

# 时点型
col_cat['time_point'] = [
    '申请周内日',
    '申请时点']

# 时长型
col_cat['time_span'] = [
    '历史贷款授信距本次申请时间',
    '首付时间']

# 布尔型
col_cat['bool'] = [
    '是否为最后一次申请记录',
    '是否为当天最后一次申请记录']

# 概率型
col_cat['prob'] = [
    '首付率',
    '利率1',
    '利率2']

# 计数型
col_cat['cnt'] = [
    '贷款期数']

### 数据清洗 - 金额型 ###
# col_cat['curr'] = [
#     '贷款年金',
#     '申请额度',
#     '信用额度',
#     '首付金额',
#     '商品价格']

# ''' 查看含有缺失值字段 '''
# for col in col_cat['curr']:
#     d3[col].value_counts(dropna=False)

# ''' 含有缺失值字段 '''
# nan_cols = [
#     '贷款年金',
#     '首付金额',
#     '商品价格'
# ]

# 数值型缺失插补均值
d3.fillna({
    '贷款年金':d3['贷款年金'].mean(),
    '首付金额':d3['首付金额'].mean(),
    '商品价格':d3['商品价格'].mean()}
     ,inplace=True)

### 数据清洗 - 时长型 ###
# col_cat['time_span'] = [
#     '历史贷款授信距本次申请时间',
#     '首付时间']

# ''' 查看含有缺失值字段 '''
# for col in col_cat['time_span']:
#     d3[col].value_counts(dropna=False)

# 数值型异常值置None
d3.loc[
    (d3['首付时间'] == 292204.0), '首付时间'] = None

# 数值型缺失均值
d3.fillna({
    '首付时间':d3['首付时间'].mean()}
     ,inplace=True)

### 数据清洗 - 概率型 ###
# col_cat['prob'] = [
#     '首付率',
#     '利率1',
#     '利率2']

# ''' 查看含有缺失值字段 '''
# for col in col_cat['prob']:
#     d3[col].value_counts(dropna=False)
    
# 数值型缺失均值
d3.fillna({
    '首付率':d3['首付率'].mean(),
    '利率1':d3['利率1'].mean(),
    '利率2':d3['利率2'].mean()}
     ,inplace=True)

### 数据清洗 - 类别型 ###
# col_cat['cat'] = [
#     '贷款类型',
#     '贷款用途',
#     '合同状态',
#     '付款方式',
#     '拒绝原因',
#     '陪同申请人',
#     '客户类型',
#     '商品类别',
#     '组合类型',
#     '产品类型',
#     '获客渠道',
#     '销售区域',
#     '行业',
#     '收益级别',
#     '产品组合']

# ''' 查看含有缺失值字段 '''
# for col in col_cat['cat']:
#     d3[col].value_counts(dropna=False)

# ''' d3 - 类别型desc '''
# d3_cat_desc = d3_feats[col_cat['cat']].astype("object").describe(include=['object'])
# print(d3_cat_desc)

# ''' 含有缺失值字段 '''
# nan_cols = [
#     '贷款类型',
#     '陪同申请人',
#     '产品组合'
# ]

### 处理含空值的离散特征，转换dummies

# d3['贷款类型'].value_counts()
# d3['陪同申请人'].value_counts()
# d3['产品组合'].value_counts()

d3_dum = pd.get_dummies(d3,dummy_na=True,columns=[
    '贷款类型',
    '陪同申请人',
    '产品组合'
])

# 空值取众数
d3_dum.loc[d3_dum['贷款类型_nan']==1, '贷款类型_0'] = 1
d3_dum.loc[d3_dum['陪同申请人_nan']==1, '陪同申请人_7.0'] = 1
d3_dum.loc[d3_dum['产品组合_nan']==1, '产品组合_2.0'] = 1

### 处理其他离散特征，有长尾分布, 小值归other类，转换dummies

# d3['贷款用途'].value_counts()

d3_dum.loc[(d3_dum['贷款用途'] != 23) & (d3_dum['贷款用途'] != 24), '贷款用途'] = 'other'

# d3['贷款用途'].value_counts()
# d3['合同状态'].value_counts()
# d3['付款方式'].value_counts()
# d3['拒绝原因'].value_counts()
# d3['客户类型'].value_counts()
# d3['商品类别'].value_counts()

d3_dum.loc[
    (d3_dum['商品类别'] != 27) & 
    (d3_dum['商品类别'] != 19) &
    (d3_dum['商品类别'] != 7) &
    (d3_dum['商品类别'] != 5) &
    (d3_dum['商品类别'] != 2) &
    (d3_dum['商品类别'] != 11), '商品类别'] = 'other'

# d3['组合类型'].value_counts()
# d3['产品类型'].value_counts()
# d3['获客渠道'].value_counts()
# d3['销售区域'].value_counts()

d3_dum['销售区域_count'] = d3.groupby('销售区域')['销售区域'].transform('count')
d3_dum.loc[d3_dum['销售区域_count'] < 5000, '销售区域'] = 'other'
# d3 = d3.drop('销售区域_count', axis=1)

# d3['行业'].value_counts()

d3_dum = pd.get_dummies(d3_dum, columns=[
    '贷款用途',
    '合同状态',
    '付款方式',
    '拒绝原因',
    '客户类型',
    '商品类别',
    '组合类型',
    '产品类型',
    '获客渠道',
    '销售区域',
    '行业'
])

### 数据清洗 - 时点型 ###

# 时点型(类别)
# col_cat['time_point'] = [
#     '申请周内日',
#     '申请时点']

# ''' 查看含有缺失值字段 '''
# for col in col_cat['time_point']:
#     d3[col].value_counts(dropna=False)

### 处理含空值的离散特征，转换dummies

# d3['申请周内日'].value_counts()
# d3['申请时点'].value_counts()

d3_dum.loc[
    (d3_dum['申请时点'] != 3) & 
    (d3_dum['申请时点'] != 4) &
    (d3_dum['申请时点'] != 5) &
    (d3_dum['申请时点'] != 6) &
    (d3_dum['申请时点'] != 7) &
    (d3_dum['申请时点'] != 8) &
    (d3_dum['申请时点'] != 9), '申请时点'] = 'other'

d3_dum = pd.get_dummies(d3_dum,columns=[
    '申请周内日',
    '申请时点'
])

### check ###

d3_dum.shape
for col in d3_dum.columns:
    print(col)
d3_dum.head()

### d3 group by ###

agg_dict = {}

# 数值型 - 金额
curr_agg = ['mean', 'max', 'min']
for col in col_cat['curr']:
    agg_dict[col] = curr_agg

# 数值型 - 时长 
time_span_agg = ['mean', 'max', 'min']
for col in col_cat['time_span']:
    agg_dict[col] = time_span_agg

# 数值型 - 概率
prob_agg = ['mean', 'max', 'min']
for col in col_cat['prob']:
    agg_dict[col] = prob_agg

# 离散型 - 类型
cat_agg = ['sum', 'mean']
for col1 in d3_dum.columns:
    for col2 in col_cat['cat']:
        if (col2 in col1):
            agg_dict[col1] = cat_agg

# 离散型 - 时点
time_point_agg = ['sum', 'mean']
for col1 in d3_dum.columns:
    for col2 in col_cat['time_point']:
        if (col2 in col1):
            agg_dict[col1] = time_point_agg



            

# print(agg_dict)

d30 = d3_dum.groupby('申请编号').agg(agg_dict)





d30.columns = pd.Series(d30.columns.tolist()).apply(pd.Series).sum(axis=1)
# d30.head

### Merge Application
d12_d30 = pd.merge(d12, d30, how='left', on='申请编号')
d12_d30.shape
d12_d30.head()

d12_d30.to_csv(f'./tmp/d12_d30.csv')
